warnings: work around most common W=123 warnings

+++ include/linux/thread_info.h
@@ -138,7 +138,7 @@ static __always_inline __must_check bool
 check_copy_size(const void *addr, size_t bytes, bool is_source)
 {
 	int sz = __compiletime_object_size(addr);
-	if (unlikely(sz >= 0 && sz < bytes)) {
+	if (unlikely(sz >= 0 && (size_t)sz < bytes)) {
 		if (!__builtin_constant_p(bytes))
 			copy_overflow(sz, bytes);
 		else if (is_source)
